scripts/pyproject.toml
[tool.poetry]
name = "upload"
version = "0.1.0"
description = ""
authors = ["Moritz Wolf <moritz.m.wolf@tum.de>"]

[tool.poetry.dependencies]
python = "^3.6"
firebase_admin = "^4.3"
requests = "^2.23"
tqdm = "^4.46"
pathlib = "^1.0"
google-cloud-firestore = "^1.9.0"
grpcio = "^1.33.1"
algoliasearch = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
